% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert data column values based on key and value}
\usage{
convert(lst, from, to, interval = FALSE)
}
\arguments{
\item{lst}{Data column to be converted}

\item{from}{Keys for dictionary as a vector}

\item{to}{Values for dictionary as a vector}

\item{interval}{Set to TRUE if converting based on partition bounds. Recommended to have length(from) = length(to)-1}
}
\description{
Convert vector based on dictionary. Can handle letter mutations if a single word is passed in. Interval is used when numeric ranges are given to from.
In this case, length(from) should be length(to) - 1 for best results
(but not necessarily necessary), and from should represent partition bounds.
Intervals are not strict. The function can also convert to different data types.
}
\examples{
convert(10:1, 1:10, letters[1:10])
}
\author{
Samuel Soon
}
