% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortable.R
\name{cortable}
\alias{cortable}
\title{make correlation table}
\usage{
cortable(df, ind_index, alternative = NULL, method = "pearson")
}
\arguments{
\item{df}{dataframe with only numeric data}

\item{ind_index}{col index of dependent variable}

\item{alternative}{vector of alternative hypotheses for each independent variable}

\item{method}{default is pearson}
}
\description{
Returns a table where columns are results of cor.test using preferred alternative and method.Create a table of correlation tests against 1 dependent variable and all other columns of dataframe
}
\author{
Samuel Soon
}
