% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountData.R
\name{getCountData}
\alias{getCountData}
\title{Summarize count data based on group and filter}
\usage{
getCountData(tab, exp, group)
}
\arguments{
\item{tab}{Data frame}

\item{exp}{Logical expression for inserting into filter function}

\item{group}{Column name for grouping, surrounded by quotations}
}
\description{
Takes a data table, filtering expression, grouping column, and outputs count data
}
\examples{
getCountData(lacrime, Age==35, "Date OCC")
}
\author{
Samuel Soon
}
