% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythagoras.R
\name{pythagoras}
\alias{pythagoras}
\title{pythagoras}
\usage{
pythagoras(a = 0, b = 0, c = 0)
}
\arguments{
\item{a}{first side}

\item{b}{second side}

\item{c}{hypotenuse side}
}
\description{
Computes third side of right triangle given two other sides
}
\examples{
pythagoras(3,4)

}
\author{
Samuel Soon
}
