% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samfunctions-package.R
\docType{package}
\name{samfunctions-package}
\alias{samfunctions}
\alias{samfunctions-package}
\title{samfunctions: Personal functions for projects}
\description{
Personal Functions
}
\author{
\strong{Maintainer}: Samuel Soon \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
