% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapirotable.R
\name{shapirotable}
\alias{shapirotable}
\title{make shapiro table}
\usage{
shapirotable(df)
}
\arguments{
\item{df}{dataframe with only numeric data}
}
\description{
Returns a table where columns are results of shapiro.test using preferred alternative and method.
}
\author{
Samuel Soon
}
