% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcol.R
\name{testcol}
\alias{testcol}
\title{make column of data frame from test results, with rounded numbers}
\usage{
testcol(l, string_data = TRUE)
}
\arguments{
\item{l}{test results given in list form}

\item{string_data}{default TRUE. Decides whether string test results are returned.}
}
\description{
Returns a vector of test parameters
}
\author{
Samuel Soon
}
